fix: pure Python 2.6 build
A recent change to the requirements file has passed a check in CI but
the build failed in a pure Py2.6 environment.  The cause is probably the
lack of support for environment markers in requirement specifiers in
older distutils.  The reason why it passed CI is probably due to Py2.7
being used to install dependencies for Py2.6.