Add mypy checks to pubtools-pyxis
The mypy checks pass in the strict mode. Some functions may return
"Any", which is due to the fact that methods of the "requests" library
may return unusual types (for example json() mat return None). __init__
method of the PyxisKrbAuth class had to be slightly rewritten so that
its argument only contained supported values.